Move edge backend to .cpp
