+++ webview-sys/webview.h
@@ -14,10 +14,9 @@ extern "C" {
 #include <stdint.h>
 #include <string.h>
 
-struct webview;
-
-typedef void (*webview_external_invoke_cb_t)(struct webview *w, const char *arg);
-typedef void (*webview_dispatch_fn)(struct webview *w, void *arg);
+typedef void* webview_t;
+typedef void (*webview_external_invoke_cb_t)(webview_t w, const char *arg);
+typedef void (*webview_dispatch_fn)(webview_t w, void *arg);
 
 enum webview_dialog_type {
   WEBVIEW_DIALOG_TYPE_OPEN = 0,
@@ -25,31 +24,31 @@ enum webview_dialog_type {
   WEBVIEW_DIALOG_TYPE_ALERT = 2
 };
 
-WEBVIEW_API int webview(const char *title, const char *url, int width,
-                        int height, int resizable);
-
-WEBVIEW_API int webview_init(struct webview *w);
-WEBVIEW_API int webview_loop(struct webview *w, int blocking);
-WEBVIEW_API int webview_eval(struct webview *w, const char *js);
-WEBVIEW_API int webview_inject_css(struct webview *w, const char *css);
-WEBVIEW_API void webview_set_title(struct webview *w, const char *title);
-WEBVIEW_API void webview_set_fullscreen(struct webview *w, int fullscreen);
-WEBVIEW_API void webview_set_color(struct webview *w, uint8_t r, uint8_t g,
+WEBVIEW_API void webview_run(webview_t w);
+WEBVIEW_API int webview_loop(webview_t w, int blocking);
+WEBVIEW_API int webview_eval(webview_t w, const char *js);
+WEBVIEW_API int webview_inject_css(webview_t w, const char *css);
+WEBVIEW_API void webview_set_title(webview_t w, const char *title);
+WEBVIEW_API void webview_set_fullscreen(webview_t w, int fullscreen);
+WEBVIEW_API void webview_set_color(webview_t w, uint8_t r, uint8_t g,
                                    uint8_t b, uint8_t a);
-WEBVIEW_API void webview_dialog(struct webview *w,
+WEBVIEW_API void webview_dialog(webview_t w,
                                 enum webview_dialog_type dlgtype, int flags,
                                 const char *title, const char *arg,
                                 char *result, size_t resultsz);
-WEBVIEW_API void webview_dispatch(struct webview *w, webview_dispatch_fn fn,
+WEBVIEW_API void webview_dispatch(webview_t w, webview_dispatch_fn fn,
                                   void *arg);
-WEBVIEW_API void webview_terminate(struct webview *w);
-WEBVIEW_API void webview_exit(struct webview *w);
+WEBVIEW_API void webview_terminate(webview_t w);
+WEBVIEW_API void webview_exit(webview_t w);
 WEBVIEW_API void webview_debug(const char *format, ...);
 WEBVIEW_API void webview_print_log(const char *s);
 
-WEBVIEW_API void* wrapper_webview_get_userdata(struct webview* w);
-WEBVIEW_API struct webview* wrapper_webview_new(const char* title, const char* url, int width, int height, int resizable, int debug, webview_external_invoke_cb_t external_invoke_cb, void* userdata);
-WEBVIEW_API void wrapper_webview_free(struct webview* w);
+WEBVIEW_API void* wrapper_webview_get_userdata(webview_t w);
+WEBVIEW_API webview_t wrapper_webview_new(const char* title, const char* url, int width, int height, int resizable, int debug, webview_external_invoke_cb_t external_invoke_cb, void* userdata);
+WEBVIEW_API void wrapper_webview_free(webview_t w);
+WEBVIEW_API void webview_destroy(webview_t w);
+
+// TODO WEBVIEW_API void webview_navigate(webview_t w, const char* url);
 
 #define WEBVIEW_DIALOG_FLAG_FILE (0 << 0)
 #define WEBVIEW_DIALOG_FLAG_DIRECTORY (1 << 0)
@@ -61,7 +60,7 @@ WEBVIEW_API void wrapper_webview_free(struct webview* w);
 
 struct webview_dispatch_arg {
   webview_dispatch_fn fn;
-  struct webview *w;
+  webview_t w;
   void *arg;
 };
